Safari-611.1.10/LayoutTests/accessibility/mac/output-element.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<output id="output">output element</output>



<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests that the output element has correct mapping.");

    if (window.accessibilityController) {

          var output = accessibilityController.accessibleElementById("output");
          shouldBe("output.role", "'AXRole: AXGroup'");
          shouldBe("output.subrole", "'AXSubrole: AXApplicationStatus'");
          shouldBe("output.roleDescription", "'AXRoleDescription: output'");
          
          shouldBe("output.title", "'AXTitle: output element'");
    }
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
