Safari-609.1.11.1/LayoutTests/accessibility/mac/output-element-expected.txt

output element
This tests that the output element has correct mapping.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS output.role is 'AXRole: AXGroup'
PASS output.subrole is 'AXSubrole: AXApplicationStatus'
PASS output.roleDescription is 'AXRoleDescription: output'
PASS output.title is 'AXTitle: output element'
PASS successfullyParsed is true

TEST COMPLETE
